move getItemFromJSONNumber to ItemParser, add docs
